Total rework of core types
