+++ cqrs-core/src/_reactor.rs
