+++ cqrs-core/src/aggregate.rs
@@ -1,268 +1,384 @@
-use std::marker::PhantomData;
+//! Aggregate related definitions.
 
-use crate::{
-    future::IntoTryFuture,
-    types::{CqrsError, EventVersion},
+#![allow(clippy::module_name_repetitions)]
+
+use std::{
+    borrow::Borrow,
+    convert::{TryFrom, TryInto as _},
+    fmt,
+    num::TryFromIntError,
+    ops,
 };
 
-/// A projected state built from a series of events.
+use async_trait::async_trait;
+
+use super::{Event, EventNumber, EventSourced, NumberedEvent};
+
+/// [DDD aggregate] that represents an isolated tree of entities, is
+/// capable of handling [`Command`]s and is always kept in a consistent state.
+///
+/// In case [`Aggregate`] is [`EventSourced`] we assume that [`Aggregate`]
+/// exists if at least one [`Event`] exists for it, so [`Aggregate`] is usable
+/// and distinguishable only after at least one [`Event`] is applied to its
+/// initial state.
+///
+/// [DDD aggregate]: https://martinfowler.com/bliki/DDD_Aggregate.html
+/// [`Command`]: super::Command
 pub trait Aggregate: Default {
     /// Type of [`Aggregate`]'s unique identifier (ID).
     type Id;
 
-    /// A static string representing the type of the aggregate.
+    /// Returns static string representing the type of this [`Aggregate`].
     ///
-    /// Note: This should effectively be a constant value, and should never change.
+    /// _Note:_ This should effectively be a constant value, and should never
+    /// change.
     fn aggregate_type() -> &'static str;
 
     /// Returns unique ID of this [`Aggregate`].
     fn id(&self) -> &Self::Id;
 }
 
-/// A command that is addressed to an [`Aggregate`].
-pub trait Command {
-    type Aggregate: Aggregate;
-
-    /// Returns ID of the [`Aggregate`] that this [`Command`] is addressed to.
-    /// `None` means that a new [`Aggregate`] should be initialized for
-    /// this [`Command`].
-    fn aggregate_id(&self) -> Option<&<Self::Aggregate as Aggregate>::Id>;
-}
-
-/// A handler of a specific [`Command`] for its [`Aggregate`].
-pub trait CommandHandler<C: Command + ?Sized> {
-    /// A context required by this [`CommandHandler`] for performing operation.
-    type Context: ?Sized;
-    type Event: Event;
+/// Source for loading snapshots of some [`Aggregate`].
+#[async_trait]
+pub trait SnapshotSource<A: Aggregate> {
+    /// Type of the shapshot loading error.
+    /// If it never fails, consider to specify [`Infallible`].
+    ///
+    /// [`Infallible`]: std::convert::Infallible
     type Err;
-    type Ok: IntoEvents<Self::Event>;
-    /// A result that this [`CommandHandler`] will return.
-    type Result: IntoTryFuture<Self::Ok, Self::Err>;
 
-    /// Handles and processes given [`Command`] for its [`Aggregate`].
-    fn handle_command(&self, cmd: &C, ctx: &Self::Context) -> Self::Result;
+    /// Loads latest stored snapshot of a given [`Aggregate`].
+    async fn load_snapshot(&self, id: &A::Id) -> Result<Option<(A, Version)>, Self::Err>;
 }
 
-/// A command that can be executed against an aggregate.
-#[deprecated]
-pub trait AggregateCommand<A: Aggregate> {
-    /// The type of event that is produced by this command.
-    type Event: AggregateEvent<A>;
+/// Sink for persisting snapshots of some [`Aggregate`].
+#[async_trait]
+pub trait SnapshotSink<A: Aggregate> {
+    /// Type of the shapshot persisting error.
+    /// If it never fails, consider to specify [`Infallible`].
+    ///
+    /// [`Infallible`]: std::convert::Infallible
+    type Err;
 
-    /// The type of the sequence of events generated when the command is executed successfully.
-    type Events: Events<ProducedEvent<A, Self>>;
-
-    /// The error type.
-    type Error: CqrsError;
-
-    /// Consumes a command, attempting to execute it against the aggregate. If the execution is successful, a sequence
-    /// of events is generated, which can be applied to the aggregate.
-    fn execute_on(self, aggregate: &A) -> Result<Self::Events, Self::Error>;
+    /// Persists [`Aggregate`]'s snapshot of a given [`Version`].
+    async fn persist_snapshot(
+        &self,
+        id: &A::Id,
+        aggregate: &A,
+        version: Version,
+    ) -> Result<(), Self::Err>;
 }
 
-/// The event type produced by this command.
-pub type ProducedEvent<A, C> = <C as AggregateCommand<A>>::Event;
-
-/// The event sequence produced by this command.
-pub type ProducedEvents<A, C> = <C as AggregateCommand<A>>::Events;
-
-/// The error produced when this command cannot be executed against an aggregate.
-pub type CommandError<A, C> = <C as AggregateCommand<A>>::Error;
-
-/// A thing that happened.
-pub trait Event {
-    /// A static description of the event.
-    fn event_type(&self) -> &'static str;
+/// [`Aggregate`] that is [`EventSourced`] and keeps track of the version of its
+/// last snapshot and the current version.
+#[derive(Clone, Copy, Debug, Default, Hash, PartialEq, Eq)]
+pub struct HydratedAggregate<A> {
+    /// Current [`Version`] of this [`Aggregate`].
+    ver: Version,
+    /// [`Version`] of last snapshot of this [`Aggregate`].
+    snapshot_ver: Option<Version>,
+    /// The [`Aggregate`] itself.
+    state: A,
 }
 
-/// A different version of the same [`Event`].
-pub trait VersionedEvent: Event {
-    /// A static version of the [`Event`].
-    fn event_version(&self) -> &'static EventVersion;
-}
+impl<A> HydratedAggregate<A> {
+    /// Creates new [`HydratedAggregate`] from a given [`Aggregate`] and its
+    /// current [`Version`].
+    #[inline]
+    pub fn from_version(agg: A, ver: Version) -> Self {
+        Self {
+            ver,
+            snapshot_ver: None,
+            state: agg,
+        }
+    }
+    /// Creates new [`HydratedAggregate`] from a given [`Aggregate`] and its
+    /// latest snapshot [`Version`].
+    #[inline]
+    pub fn from_snapshot(agg: A, ver: Version) -> Self {
+        Self {
+            ver,
+            snapshot_ver: Some(ver),
+            state: agg,
+        }
+    }
 
-/// A state that can be calculated by applying specified [`Event`].
-pub trait EventSourced<E: Event> {
-    /// Applies given [`Event`] to the current state.
-    fn apply_event(&mut self, event: &E);
-}
-
-/// An event that can be applied to an aggregate.
-#[deprecated]
-pub trait AggregateEvent<A: Aggregate>: Event {
-    /// Consumes the event, applying its effects to the aggregate.
-    fn apply_to(self, aggregate: &mut A);
-}
-
-/// An iterable and sliceable list of events.
-pub trait Events<E>: IntoIterator<Item = E> + AsRef<[E]>
-where
-    E: Event,
-{
-}
-
-impl<T, E> Events<E> for T
-where
-    T: IntoIterator<Item = E> + AsRef<[E]>,
-    E: Event,
-{
-}
-
-/// An event that can be serialized to a buffer.
-pub trait SerializableEvent: Event {
-    /// The error type.
-    type Error: CqrsError;
-
-    /// Serializes the event to the given buffer.
-    fn serialize_event_to_buffer(&self, buffer: &mut Vec<u8>) -> Result<(), Self::Error>;
-}
-
-/// An event that can be deserialized from a buffer.
-pub trait DeserializableEvent: Event + Sized {
-    /// The error type.
-    type Error: CqrsError;
-
-    /// Deserializes an event from the provided buffer, with prior knowledge about the event's type.
-    fn deserialize_event_from_buffer(
-        data: &[u8],
-        event_type: &str,
-    ) -> Result<Option<Self>, Self::Error>;
-}
-
-pub trait IntoEvents<E> {
-    type Iter: AsRef<[E]>;
-
-    fn into_events(self) -> Self::Iter;
-}
-
-impl<E> IntoEvents<E> for () {
-    type Iter = [E; 0];
-
+    /// Returns ID of this [`Aggregate`].
     #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        []
+    pub fn id(&self) -> &A::Id
+    where
+        A: Aggregate,
+    {
+        self.state.id()
+    }
+
+    /// Returns the current [`Version`] of this [`Aggregate`].
+    #[inline(always)]
+    pub fn version(&self) -> Version {
+        self.ver
+    }
+
+    /// Returns [`Version`] of the latest snapshot for this [`Aggregate`].
+    #[inline(always)]
+    pub fn snapshot_version(&self) -> Option<Version> {
+        self.snapshot_ver
+    }
+
+    /// Sets [`Version`] of the latest snapshot for this [`Aggregate`].
+    #[inline(always)]
+    pub fn set_snapshot_version(&mut self, new: Version) {
+        self.snapshot_ver = Some(new);
+    }
+
+    /// Returns the inner [`Aggregate`] itself.
+    #[inline(always)]
+    pub fn state(&self) -> &A {
+        &self.state
+    }
+
+    /// Applies given [`NumberedEvent`] to this [`Aggregate`] renewing its
+    /// [`Version`] with [`EventNumber`] of the [`NumberedEvent`].
+    #[inline]
+    pub fn apply<'a, E, R>(&mut self, event: R)
+    where
+        R: Into<NumberedEvent<&'a E>>,
+        E: Event + 'a,
+        A: EventSourced<E>,
+    {
+        let e = event.into();
+        self.state.apply_event(e.data);
+        self.ver = e.num.into();
+    }
+
+    /// Applies given [`NumberedEvent`]s to this [`Aggregate`] renewing its
+    /// [`Version`] appropriately.
+    #[inline]
+    pub fn apply_events<'a, E, R, I>(&mut self, events: I)
+    where
+        I: IntoIterator<Item = R>,
+        R: Into<NumberedEvent<&'a E>>,
+        E: Event + 'a,
+        A: EventSourced<E>,
+    {
+        for event in events {
+            self.apply(event);
+        }
     }
 }
 
-impl<E, A> IntoEvents<E> for (A,)
-where
-    E: From<A>,
-{
-    type Iter = [E; 1];
-
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        [self.0.into()]
+impl<A> AsRef<A> for HydratedAggregate<A> {
+    #[inline]
+    fn as_ref(&self) -> &A {
+        &self.state
     }
 }
 
-impl<E, A, B> IntoEvents<E> for (A, B)
-where
-    E: From<A> + From<B>,
-{
-    type Iter = [E; 2];
-
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        [self.0.into(), self.1.into()]
+impl<A> Borrow<A> for HydratedAggregate<A> {
+    #[inline]
+    fn borrow(&self) -> &A {
+        &self.state
     }
 }
 
-impl<E, A, B, C> IntoEvents<E> for (A, B, C)
-where
-    E: From<A> + From<B> + From<C>,
-{
-    type Iter = [E; 3];
+/// Version of an [`Aggregate`].
+#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
+pub enum Version {
+    /// Version has no [`Event`]s applied to it.
+    Initial,
+    /// Version of the last [`Event`] applied to an [`Aggregate`].
+    Number(EventNumber),
+}
 
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        [self.0.into(), self.1.into(), self.2.into()]
+impl Default for Version {
+    #[inline]
+    fn default() -> Self {
+        Version::Initial
     }
 }
 
-impl<E, A, B, C, D> IntoEvents<E> for (A, B, C, D)
-where
-    E: From<A> + From<B> + From<C> + From<D>,
-{
-    type Iter = [E; 4];
+impl Version {
+    /// Creates new [`Version`] from a number.
+    ///
+    /// The number `0` gets interpreted as being [`Version::Initial`], while any
+    /// other number is interpreted as the latest [`EventNumber`] applied.
+    #[inline]
+    pub fn new<N: Into<u128>>(number: N) -> Self {
+        EventNumber::new(number).map_or(Version::Initial, Version::Number)
+    }
 
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        [self.0.into(), self.1.into(), self.2.into(), self.3.into()]
+    /// Increments [`Version`] number to the next in sequence.
+    #[inline]
+    pub fn incr(&mut self) {
+        match *self {
+            Version::Initial => *self = Version::Number(EventNumber::MIN_VALUE),
+            Version::Number(ref mut en) => en.incr(),
+        }
+    }
+
+    /// Returns next [`EventNumber`] in a sequence.
+    #[inline]
+    pub fn next_event(self) -> EventNumber {
+        match self {
+            Version::Initial => EventNumber::MIN_VALUE,
+            Version::Number(mut en) => {
+                en.incr();
+                en
+            }
+        }
+    }
+
+    /// Returns [`Version`] number as [`EventNumber`], returning [`None`] if the
+    /// current [`Version`] is [`Version::Initial`].
+    #[inline]
+    pub fn event_number(self) -> Option<EventNumber> {
+        match self {
+            Version::Initial => None,
+            Version::Number(en) => Some(en),
+        }
     }
 }
 
-impl<E> IntoEvents<E> for Vec<E> {
-    type Iter = Self;
-
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self
+impl fmt::Display for Version {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match *self {
+            Version::Initial => f.write_str("initial"),
+            Version::Number(ref event_number) => event_number.fmt(f),
+        }
     }
 }
 
-impl<E> IntoEvents<E> for [E; 0] {
-    type Iter = Self;
-
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self
+impl From<EventNumber> for Version {
+    #[inline]
+    fn from(n: EventNumber) -> Self {
+        Version::Number(n)
     }
 }
 
-impl<E> IntoEvents<E> for [E; 1] {
-    type Iter = Self;
+macro_rules! impl_into_version_for {
+    ($t:ty) => {
+        impl From<$t> for Version {
+            #[inline]
+            fn from(n: $t) -> Self {
+                Self::new(n)
+            }
+        }
+    };
+}
+impl_into_version_for!(u8);
+impl_into_version_for!(u16);
+impl_into_version_for!(u32);
+impl_into_version_for!(u64);
+impl_into_version_for!(u128);
 
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self
+macro_rules! impl_try_into_version_for {
+    ($t:ty) => {
+        impl TryFrom<$t> for Version {
+            type Error = TryFromIntError;
+
+            #[inline]
+            fn try_from(n: $t) -> Result<Self, Self::Error> {
+                match n {
+                    0 => Ok(Version::Initial),
+                    _ => n.try_into(),
+                }
+            }
+        }
+    };
+}
+impl_try_into_version_for!(i8);
+impl_try_into_version_for!(i16);
+impl_try_into_version_for!(i32);
+impl_try_into_version_for!(i64);
+impl_try_into_version_for!(i128);
+impl_try_into_version_for!(usize);
+impl_try_into_version_for!(isize);
+
+impl From<Version> for u128 {
+    #[inline]
+    fn from(v: Version) -> Self {
+        match v {
+            Version::Initial => 0,
+            Version::Number(n) => n.into(),
+        }
     }
 }
 
-impl<E> IntoEvents<E> for [E; 2] {
-    type Iter = Self;
+macro_rules! impl_try_from_version_for {
+    ($t:ty) => {
+        impl TryFrom<Version> for $t {
+            type Error = TryFromIntError;
 
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self
+            #[inline]
+            fn try_from(v: Version) -> Result<Self, Self::Error> {
+                match v {
+                    Version::Initial => Ok(0),
+                    Version::Number(n) => n.try_into(),
+                }
+            }
+        }
+    };
+}
+impl_try_from_version_for!(u8);
+impl_try_from_version_for!(i8);
+impl_try_from_version_for!(u16);
+impl_try_from_version_for!(i16);
+impl_try_from_version_for!(u32);
+impl_try_from_version_for!(i32);
+impl_try_from_version_for!(u64);
+impl_try_from_version_for!(i64);
+impl_try_from_version_for!(i128);
+impl_try_from_version_for!(usize);
+impl_try_from_version_for!(isize);
+
+impl ops::Sub for Version {
+    type Output = i128;
+
+    fn sub(self, rhs: Self) -> Self::Output {
+        i128::try_from(self).unwrap() - i128::try_from(rhs).unwrap()
     }
 }
 
-impl<E> IntoEvents<E> for [E; 3] {
-    type Iter = Self;
+/// Recommendation on whether or not a snapshot of an [`Aggregate`] should be
+/// persisted.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub enum SnapshotRecommendation {
+    /// Snapshot should be taken.
+    ShouldSnapshot,
+    /// Snapshot should not be taken.
+    DoNotSnapshot,
+}
 
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self
+/// Strategy determining when a snapshot of an [`Aggregate`] should be taken.
+pub trait SnapshotStrategy {
+    /// Gives the [`SnapshotRecommendation`] on whether or not to perform
+    /// a snapshot for an [`Aggregate`].
+    fn snapshot_recommendation(
+        &self,
+        ver: Version,
+        last_snapshot_ver: Option<Version>,
+    ) -> SnapshotRecommendation;
+}
+
+/// [`SnapshotStrategy`] that will never recommend taking a snapshot.
+#[derive(Clone, Copy, Debug, Default, Hash, PartialEq, Eq)]
+pub struct NeverSnapshot;
+
+impl SnapshotStrategy for NeverSnapshot {
+    /// Always returns [`SnapshotRecommendation::DoNotSnapshot`].
+    #[inline]
+    fn snapshot_recommendation(&self, _: Version, _: Option<Version>) -> SnapshotRecommendation {
+        SnapshotRecommendation::DoNotSnapshot
     }
 }
 
-impl<E> IntoEvents<E> for [E; 4] {
-    type Iter = Self;
+/// [`SnapshotStrategy`] that will always recommend taking a snapshot.
+#[derive(Clone, Copy, Debug, Default, Hash, PartialEq, Eq)]
+pub struct AlwaysSnapshot;
 
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self
-    }
-}
-
-impl<E, T> IntoEvents<E> for EventsRef<T>
-where
-    T: AsRef<[E]>,
-{
-    type Iter = T;
-
-    #[inline(always)]
-    fn into_events(self) -> Self::Iter {
-        self.0
-    }
-}
-
-pub struct EventsRef<T>(pub T);
-
-impl<T> From<T> for EventsRef<T> {
-    #[inline(always)]
-    fn from(v: T) -> Self {
-        Self(v)
+impl SnapshotStrategy for AlwaysSnapshot {
+    /// Always returns [`SnapshotRecommendation::ShouldSnapshot`].
+    #[inline]
+    fn snapshot_recommendation(&self, _: Version, _: Option<Version>) -> SnapshotRecommendation {
+        SnapshotRecommendation::ShouldSnapshot
     }
 }
